+++ src/pipx/util.py
@@ -114,17 +114,19 @@ def run_pypackage_bin(bin_path: Path, args: List[str]) -> NoReturn:
 
 if WINDOWS:
 
-    def get_venv_paths(root: Path) -> Tuple[Path, Path]:
+    def get_venv_paths(root: Path) -> Tuple[Path, Path, Path]:
         bin_path = root / "Scripts" if not MINGW else root / "bin"
         python_path = bin_path / "python.exe"
-        return bin_path, python_path
+        man_path = root / "share" / "man"
+        return bin_path, python_path, man_path
 
 else:
 
-    def get_venv_paths(root: Path) -> Tuple[Path, Path]:
+    def get_venv_paths(root: Path) -> Tuple[Path, Path, Path]:
         bin_path = root / "bin"
         python_path = bin_path / "python"
-        return bin_path, python_path
+        man_path = root / "share" / "man"
+        return bin_path, python_path, man_path
 
 
 def get_site_packages(python: Path) -> Path: