+++ tests/test_environment.py
@@ -6,6 +6,7 @@ def test_cli(monkeypatch, capsys):
     captured = capsys.readouterr()
     assert "PIPX_HOME" in captured.out
     assert "PIPX_BIN_DIR" in captured.out
+    assert "PIPX_MAN_DIR" in captured.out
     assert "PIPX_SHARED_LIBS" in captured.out
     assert "PIPX_LOCAL_VENVS" in captured.out
     assert "PIPX_LOG_DIR" in captured.out
@@ -19,6 +20,7 @@ def test_cli(monkeypatch, capsys):
 def test_cli_with_args(monkeypatch, capsys):
     assert not run_pipx_cli(["environment", "--value", "PIPX_HOME"])
     assert not run_pipx_cli(["environment", "--value", "PIPX_BIN_DIR"])
+    assert not run_pipx_cli(["environment", "--value", "PIPX_MAN_DIR"])
     assert not run_pipx_cli(["environment", "--value", "PIPX_SHARED_LIBS"])
     assert not run_pipx_cli(["environment", "--value", "PIPX_LOCAL_VENVS"])
     assert not run_pipx_cli(["environment", "--value", "PIPX_LOG_DIR"])