+++ tests/test_install.py
@@ -136,7 +136,7 @@ def test_name_tricky_characters(
 
 
 def test_extra(pipx_temp_env, capsys):
-    assert not run_pipx_cli(["install", "nox[tox_to_nox]==2020.8.22", "--include-deps"])
+    assert not run_pipx_cli(["install", "nox[tox_to_nox]==2023.4.22", "--include-deps"])
     captured = capsys.readouterr()
     assert f"- {app_name('tox')}\n" in captured.out
 
@@ -147,6 +147,7 @@ def test_install_local_extra(pipx_temp_env, capsys):
     )
     captured = capsys.readouterr()
     assert f"- {app_name('pycowsay')}\n" in captured.out
+    assert f"- {Path('man6/pycowsay.6')}\n" in captured.out
 
 
 def test_path_warning(pipx_temp_env, capsys, monkeypatch, caplog):
@@ -177,12 +178,39 @@ def test_existing_symlink_points_to_existing_wrong_location_warning(
     assert "is not on your PATH environment variable" not in captured.err
 
 
+def test_existing_man_page_symlink_points_to_existing_wrong_location_warning(
+    pipx_temp_env, caplog, capsys
+):
+    if sys.platform.startswith("win"):
+        pytest.skip("pipx does not use symlinks on Windows")
+
+    (constants.LOCAL_MAN_DIR / "man6").mkdir(exist_ok=True, parents=True)
+    (constants.LOCAL_MAN_DIR / "man6" / "pycowsay.6").symlink_to(os.devnull)
+    assert not run_pipx_cli(["install", "pycowsay"])
+    captured = capsys.readouterr()
+    assert "File exists at" in unwrap_log_text(caplog.text)
+    assert "symlink missing or pointing to unexpected location" in captured.out
+
+
 def test_existing_symlink_points_to_nothing(pipx_temp_env, capsys):
     if sys.platform.startswith("win"):
         pytest.skip("pipx does not use symlinks on Windows")
 
     constants.LOCAL_BIN_DIR.mkdir(exist_ok=True, parents=True)
     (constants.LOCAL_BIN_DIR / "pycowsay").symlink_to("/asdf/jkl")
+    assert not run_pipx_cli(["install", "pycowsay"])
+    captured = capsys.readouterr()
+    # pipx should realize the symlink points to nothing and replace it,
+    # so no warning should be present
+    assert "symlink missing or pointing to unexpected location" not in captured.out
+
+
+def test_existing_man_page_symlink_points_to_nothing(pipx_temp_env, capsys):
+    if sys.platform.startswith("win"):
+        pytest.skip("pipx does not use symlinks on Windows")
+
+    (constants.LOCAL_MAN_DIR / "man6").mkdir(exist_ok=True, parents=True)
+    (constants.LOCAL_MAN_DIR / "man6" / "pycowsay.6").symlink_to("/asdf/jkl")
     assert not run_pipx_cli(["install", "pycowsay"])
     captured = capsys.readouterr()
     # pipx should realize the symlink points to nothing and replace it,
@@ -237,6 +265,13 @@ def test_install_suffix(pipx_temp_env, capsys):
 
     assert (constants.LOCAL_BIN_DIR / name_a).exists()
     assert (constants.LOCAL_BIN_DIR / name_b).exists()
+
+
+def test_man_page_install(pipx_temp_env, capsys):
+    assert not run_pipx_cli(["install", "pycowsay"])
+    captured = capsys.readouterr()
+    assert f"- {Path('man6/pycowsay.6')}" in captured.out
+    assert (constants.LOCAL_MAN_DIR / "man6" / "pycowsay.6").exists()
 
 
 def test_install_pip_failure(pipx_temp_env, capsys):