+++ tests/test_run.py
@@ -45,7 +45,7 @@ def run_pipx_cli_exit(pipx_cmd_list, assert_exit=None):
 
 
 @pytest.mark.parametrize(
-    "package_name", ["pycowsay", "pycowsay==0.0.0.1", "pycowsay>=0.0.0.1"]
+    "package_name", ["pycowsay", "pycowsay==0.0.0.2", "pycowsay>=0.0.0.2"]
 )
 @mock.patch("os.execvpe", new=execvpe_mock)
 def test_simple_run(pipx_temp_env, monkeypatch, capsys, package_name):
@@ -209,7 +209,7 @@ def test_run_with_requirements(caplog, pipx_temp_env, tmp_path):
         textwrap.dedent(
             f"""
                 # /// pyproject
-                # run.requirements = ["requests==2.28.1"]
+                # run.requirements = ["requests==2.31.0"]
                 # ///
 
                 # Check requests can be imported
@@ -224,7 +224,7 @@ def test_run_with_requirements(caplog, pipx_temp_env, tmp_path):
         encoding="utf-8",
     )
     run_pipx_cli_exit(["run", script.as_uri()])
-    assert out.read_text() == "2.28.1"
+    assert out.read_text() == "2.31.0"
 
 
 @mock.patch("os.execvpe", new=execvpe_mock)